Windows install script should indicate installer failure.

+++ installers/install.ps1
@@ -147,6 +147,9 @@ $PSDefaultParameterValues['*:Encoding'] = 'utf8'
 
 $env:ACTIVESTATE_SESSION_TOKEN = $script:SESSION_TOKEN_VALUE
 & $exePath $args --source-installer="install.ps1"
+if ( !$? ) {
+  return 1
+}
 if (Test-Path env:ACTIVESTATE_SESSION_TOKEN)
 {
     Remove-Item Env:\ACTIVESTATE_SESSION_TOKEN
